supporting f1x-bench options

+++ tools/F1XBench.cpp
@@ -52,11 +52,8 @@ using std::string;
 
 
 enum class ExperimentStatus {
-  SUCCESS, FAILURE, TIMEOUT
+  SUCCESS, FAILURE, ERROR, TIMEOUT
 };
-
-
-const unsigned TIMEOUT_STATUS = 124;
 
 
 std::string statusToString(ExperimentStatus status) {
@@ -65,6 +62,8 @@ std::string statusToString(ExperimentStatus status) {
     return "SUCCESS";
   case ExperimentStatus::FAILURE:
     return "FAILURE";
+  case ExperimentStatus::ERROR:
+    return "ERROR";
   case ExperimentStatus::TIMEOUT:
     return "TIMEOUT";
   }
@@ -99,48 +98,68 @@ std::string formatMilliseconds(unsigned long milliseconds) {
       out << days << "d " << hours << "h " << minutes << "m " << seconds << "s";
     }
   } else {
-    out << "less than a second";
+    out << "<1s";
   }
   return out.str();
 }
 
-ExperimentStatus experiment(std::string defectId, 
-                            fs::path root, 
-                            fs::path output, 
-                            unsigned timeout,
-                            bool verbose) {
 
-  InEnvironment env({ {"CC", "f1x-cc"},
-                      {"F1X_BENCH_ROOT", root.string()},
-                      {"F1X_BENCH_OUTPUT", output.string()} });
-
-  if (fs::exists(output)) {
-    fs::create_directory(output);
-  }
-
-  fs::path benchmarkFile = root / "benchmark.json";
-  fs::path testsFile = root / "tests.json";
-  fs::path patchFile = output / (defectId + ".patch");
-  fs::path defectDir(defectId);
-
-  json::Document benchmark;
-  {
-    fs::ifstream ifs(benchmarkFile);
-    json::IStreamWrapper isw(ifs);
-    benchmark.ParseStream(isw);
-  }
-
+bool fetch(std::string defectId, const json::Document &benchmark) {
   auto defect = benchmark.GetObject()[defectId.c_str()].GetObject();
-
   std::string fetch = defect["fetch"].GetString();
+  std::stringstream fetchCmd;
+  fetchCmd << fetch << " >/dev/null 2>&1";
+  BOOST_LOG_TRIVIAL(debug) << defectId << "/fetch: " << fetchCmd.str();
+  unsigned long fetchStatus = std::system(fetchCmd.str().c_str());
+  if (WEXITSTATUS(fetchStatus) != 0) {
+    return false;
+  }
+  return true;
+}
+
+
+bool setUp(std::string defectId, const json::Document &benchmark) {
+  auto defect = benchmark.GetObject()[defectId.c_str()].GetObject();
   std::string setUp;
   if (defect.HasMember("set-up")) {
     setUp = defect["set-up"].GetString();
+  } else {
+    BOOST_LOG_TRIVIAL(debug) << defectId << "/set-up is undefined";
+    return true;
   }
+  std::stringstream setUpCmd;
+  setUpCmd << setUp  << " >/dev/null 2>&1";
+  BOOST_LOG_TRIVIAL(debug) << defectId << "/set-up: " << setUpCmd.str();
+  unsigned long setUpStatus = std::system(setUpCmd.str().c_str());
+  if (WEXITSTATUS(setUpStatus) != 0) {
+    return false;
+  }
+  return true;
+}
+
+
+bool tearDown(std::string defectId, const json::Document &benchmark) {
+  auto defect = benchmark.GetObject()[defectId.c_str()].GetObject();
   std::string tearDown;
-  if (defect.HasMember("tear-down")) {
-    tearDown = defect["tear-down"].GetString();
+  if (defect.HasMember("set-up")) {
+    tearDown = defect["set-up"].GetString();
+  } else {
+    BOOST_LOG_TRIVIAL(debug) << defectId << "/tear-down is undefined";
+    return true;
   }
+  std::stringstream tearDownCmd;
+  tearDownCmd << tearDown  << " >/dev/null 2>&1";
+  BOOST_LOG_TRIVIAL(debug) << defectId << "/tear-down: " << tearDownCmd.str();
+  unsigned long tearDownStatus = std::system(tearDownCmd.str().c_str());
+  if (WEXITSTATUS(tearDownStatus) != 0) {
+    return false;
+  }
+  return true;
+}
+
+
+std::string cmd(std::string defectId, const json::Document &benchmark, const json::Document &alltests, unsigned timeout, fs::path output, bool verbose) {
+  auto defect = benchmark.GetObject()[defectId.c_str()].GetObject();
   std::string source = defect["source"].GetString();
   vector<std::string> files;
   for (json::SizeType i = 0; i < defect["files"].GetArray().Size(); i++)
@@ -151,14 +170,6 @@ ExperimentStatus experiment(std::string defectId,
   }
   unsigned testTimeout = defect["test-timeout"].GetUint();
   std::string driver = defect["driver"].GetString();
-
-  json::Document alltests;
-  {
-    fs::ifstream ifs(testsFile);
-    json::IStreamWrapper isw(ifs);
-    alltests.ParseStream(isw);
-  }
-
   auto tests = alltests.GetObject()[defectId.c_str()].GetObject();
   vector<std::string> negative;
   for (json::SizeType i = 0; i < tests["negative"].GetArray().Size(); i++)
@@ -166,24 +177,6 @@ ExperimentStatus experiment(std::string defectId,
   vector<std::string> positive;
   for (json::SizeType i = 0; i < tests["positive"].GetArray().Size(); i++)
     positive.push_back(tests["positive"].GetArray()[i].GetString());
-
-  std::stringstream fetchCmd;
-  fetchCmd << fetch << " >/dev/null 2>&1";
-  BOOST_LOG_TRIVIAL(debug) << defectId << "/fetch: " << fetchCmd.str();
-  unsigned long fetchStatus = std::system(fetchCmd.str().c_str());
-  if (WEXITSTATUS(fetchStatus) != 0) {
-    BOOST_LOG_TRIVIAL(warning) << defectId << "/fetch returned non-zero exit code";
-  }
-
-  if (!setUp.empty()) {
-    std::stringstream setUpCmd;
-    setUpCmd << setUp << " >/dev/null 2>&1";
-    BOOST_LOG_TRIVIAL(debug) << defectId << "/set-up: " << setUpCmd.str();
-    unsigned long setUpStatus = std::system(setUpCmd.str().c_str());
-    if (WEXITSTATUS(setUpStatus) != 0) {
-      BOOST_LOG_TRIVIAL(warning) << defectId << "/set-up returned non-zero exit code";
-    }
-  }
 
   std::stringstream f1xCmd;
   f1xCmd << "timeout " << timeout << "s" << " f1x " << source;
@@ -203,6 +196,7 @@ ExperimentStatus experiment(std::string defectId,
   if (!build.empty()) {
     f1xCmd << " --build " << build;
   }
+  fs::path patchFile = output / (defectId + ".patch");
   f1xCmd << " --output " << patchFile.string();
   if (verbose) {
     f1xCmd << " >&2";
@@ -211,35 +205,96 @@ ExperimentStatus experiment(std::string defectId,
     f1xCmd << " >/dev/null 2>&1";
   }
 
-  BOOST_LOG_TRIVIAL(debug) << defectId << "/run: " << f1xCmd.str();
-  
-  unsigned long f1xStatus = std::system(f1xCmd.str().c_str());
+  return f1xCmd.str();
+}
 
-  ExperimentStatus result;
+
+std::pair<ExperimentStatus, unsigned long> run(std::string defectId, const json::Document &benchmark, const json::Document &alltests, unsigned timeout, fs::path output, bool verbose) {
+  std::string f1xCmd = cmd(defectId, benchmark, alltests, timeout, output, verbose);
+  BOOST_LOG_TRIVIAL(debug) << defectId << "/run: " << f1xCmd;
+
+  std::chrono::steady_clock::time_point begin = std::chrono::steady_clock::now();
+  unsigned long f1xStatus = std::system(f1xCmd.c_str());
+  std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();
   
+  unsigned int defectTime = std::chrono::duration_cast<std::chrono::milliseconds>(end - begin).count();
+  ExperimentStatus result;
   if (WEXITSTATUS(f1xStatus) == 0) {
     result = ExperimentStatus::SUCCESS;
-  } else if (WEXITSTATUS(f1xStatus) == TIMEOUT_STATUS) {
+  } else if (WEXITSTATUS(f1xStatus) == FAILURE_EXIT_CODE) {
+    result = ExperimentStatus::FAILURE;
+  } else if (WEXITSTATUS(f1xStatus) == TIMEOUT_EXIT_CODE) {
     result = ExperimentStatus::TIMEOUT;
   } else {
-    result = ExperimentStatus::FAILURE;
+    result = ExperimentStatus::ERROR;
   }
 
-  if (!tearDown.empty()) {
-    std::stringstream tearDownCmd;
-    tearDownCmd << tearDown << " >/dev/null 2>&1";
-    BOOST_LOG_TRIVIAL(debug) << defectId << "/tear-down: " << tearDownCmd.str();
-    unsigned long tearDownStatus = std::system(tearDownCmd.str().c_str());
-    if (WEXITSTATUS(tearDownStatus) != 0) {
+  return std::make_pair(result, defectTime);
+}
+
+
+struct ExperimentPlan {
+  bool fetch;
+  bool setUp;
+  bool run;
+  bool tearDown;
+  bool remove;
+};
+
+
+std::pair<ExperimentStatus, unsigned long> experiment(std::string defectId, 
+                                                      fs::path root, 
+                                                      fs::path output,
+                                                      const json::Document &benchmark,
+                                                      const json::Document &alltests,
+                                                      unsigned timeout,
+                                                      bool verbose,
+                                                      ExperimentPlan plan) {
+
+  InEnvironment env({ {"CC", "f1x-cc"},
+                      {"F1X_BENCH_ROOT", root.string()},
+                      {"F1X_BENCH_OUTPUT", output.string()} });
+
+  if (fs::exists(output)) {
+    fs::create_directory(output);
+  }
+
+  fs::path defectDir(defectId);
+
+  if (plan.fetch) {
+    bool fetchSuccess = fetch(defectId, benchmark);
+    if (! fetchSuccess) {
+      BOOST_LOG_TRIVIAL(warning) << defectId << "/fetch returned non-zero exit code";
+    }
+  }
+
+  if (plan.setUp) {
+    bool setUpSuccess = setUp(defectId, benchmark);
+    if (! setUpSuccess) {
+      BOOST_LOG_TRIVIAL(warning) << defectId << "/set-up returned non-zero exit code";
+    }
+  }
+
+  std::pair<ExperimentStatus, unsigned long> result;
+  
+  if (plan.run) {
+    result = run(defectId, benchmark, alltests, timeout, output, verbose);
+  }
+
+  if (plan.tearDown) {
+    bool tearDownSuccess = tearDown(defectId, benchmark);
+    if (! tearDownSuccess) {
       BOOST_LOG_TRIVIAL(warning) << defectId << "/tear-down returned non-zero exit code";
     }
   }
 
-  BOOST_LOG_TRIVIAL(debug) << "removing " << defectDir;
-  if (fs::exists(defectDir)) {
-    fs::remove_all(defectDir);
+  if (plan.remove) {
+    BOOST_LOG_TRIVIAL(debug) << "removing " << defectDir;
+    if (fs::exists(defectDir)) {
+      fs::remove_all(defectDir);
+    }
   }
- 
+
   return result;
 }
 
@@ -301,6 +356,56 @@ int main (int argc, char *argv[]) {
 
   initializeTrivialLogger(verbose);
 
+  bool fetchOnly = false;
+  if (vm.count("fetch")) {
+    if (!vm.count("defect")) {
+      BOOST_LOG_TRIVIAL(error) << "defect is not specified";
+      return 0;
+    } else {
+      fetchOnly = true;
+    }
+  }
+
+  bool setUpOnly = false;
+  if (vm.count("set-up")) {
+    if (!vm.count("defect")) {
+      BOOST_LOG_TRIVIAL(error) << "defect is not specified";
+      return 0;
+    } else {
+      setUpOnly = true;
+    }
+  }
+
+  bool runOnly = false;
+  if (vm.count("run")) {
+    if (!vm.count("defect")) {
+      BOOST_LOG_TRIVIAL(error) << "defect is not specified";
+      return 0;
+    } else {
+      runOnly = true;
+    }
+  }
+
+  bool tearDownOnly = false;
+  if (vm.count("tear-down")) {
+    if (!vm.count("defect")) {
+      BOOST_LOG_TRIVIAL(error) << "defect is not specified";
+      return 0;
+    } else {
+      tearDownOnly = true;
+    }
+  }
+
+  bool cmdOnly = false;
+  if (vm.count("cmd")) {
+    if (!vm.count("defect")) {
+      BOOST_LOG_TRIVIAL(error) << "defect is not specified";
+      return 0;
+    } else {
+      cmdOnly = true;
+    }
+  }
+
   fs::path root(fs::current_path());
   if (vm.count("root")) {
     fs::path specifiedRoot(vm["root"].as<string>());
@@ -347,19 +452,35 @@ int main (int argc, char *argv[]) {
     fs::create_directory(output);
   }
 
+  fs::path benchmarkFile = root / "benchmark.json";
+  fs::path testsFile = root / "tests.json";
+
+  json::Document benchmark;
+  {
+    fs::ifstream ifs(benchmarkFile);
+    json::IStreamWrapper isw(ifs);
+    benchmark.ParseStream(isw);
+  }
+
+  json::Document alltests;
+  {
+    fs::ifstream ifs(testsFile);
+    json::IStreamWrapper isw(ifs);
+    alltests.ParseStream(isw);
+  }
+
+
+
   vector<std::string> defects;
 
   if (vm.count("defect")) {
+    if (cmdOnly) {
+      std::string f1xCmd = cmd(vm["defect"].as<string>(), benchmark, alltests, timeout, output, verbose);
+      std::cout << f1xCmd << std::endl;
+      return 0;
+    }
     defects.push_back(vm["defect"].as<string>());
   } else {
-    fs::path benchmarkFile = root / "benchmark.json";
-
-    json::Document benchmark;
-    {
-      fs::ifstream ifs(benchmarkFile);
-      json::IStreamWrapper isw(ifs);
-      benchmark.ParseStream(isw);
-    }
     for (json::Value::ConstMemberIterator itr = benchmark.MemberBegin(); itr != benchmark.MemberEnd(); ++itr) {
       defects.push_back(itr->name.GetString());
     }
@@ -367,31 +488,44 @@ int main (int argc, char *argv[]) {
 
   unsigned long totalTime = 0;
   unsigned long totalRepaired = 0;
+  unsigned long totalErrors = 0;
+  unsigned long totalTimeouts = 0;
 
   for (auto &defect : defects) {
     unsigned long defectTime = 0;
 
-    std::chrono::steady_clock::time_point begin = std::chrono::steady_clock::now();
-    
-    ExperimentStatus status = experiment(defect, root, output, timeout, verbose);
-  
-    std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();
+    ExperimentPlan plan{ !(setUpOnly || runOnly || tearDownOnly),
+                         !(fetchOnly || runOnly || tearDownOnly),
+                         !(fetchOnly || setUpOnly || tearDownOnly),
+                         !(fetchOnly || setUpOnly || runOnly),
+                         !(fetchOnly || setUpOnly || runOnly || tearDownOnly) };
 
-    defectTime = std::chrono::duration_cast<std::chrono::milliseconds>(end - begin).count();
-    totalTime += defectTime;
+    std::pair<ExperimentStatus, unsigned long> result = experiment(defect, root, output, benchmark, alltests, timeout, verbose, plan);
+    
+    if (fetchOnly || setUpOnly || tearDownOnly)
+      continue;
+      
+    totalTime += result.second;
 
     BOOST_LOG_TRIVIAL(info) << defect 
-                            << "\t" << statusToString(status)
-                            << "\t" << formatMilliseconds(defectTime);
+                            << "\t" << statusToString(result.first)
+                            << "\t" << formatMilliseconds(result.second);
 
-    if (status == ExperimentStatus::SUCCESS) {
+    if (result.first == ExperimentStatus::SUCCESS) {
       totalRepaired++;
     }
+    if (result.first == ExperimentStatus::ERROR) {
+      totalErrors++;
+    }
+    if (result.first == ExperimentStatus::TIMEOUT) {
+      totalTimeouts++;
+    }
+    
   }
 
-  if (defects.size() > 1) {
-    BOOST_LOG_TRIVIAL(info) << "total time: " << formatMilliseconds(totalTime);
-    BOOST_LOG_TRIVIAL(info) << "total repaired: " << totalRepaired << "/" << defects.size();
+  if (defects.size() > 1 && !(fetchOnly || setUpOnly || tearDownOnly)) {
+    BOOST_LOG_TRIVIAL(info) << "time: " << formatMilliseconds(totalTime);
+    BOOST_LOG_TRIVIAL(info) << "repaired/errors/timeouts/total: " << totalRepaired << "/" << totalErrors << "/" << totalTimeouts << "/" << defects.size();
   }
 
   return 0;
